+++ src/lib.rs
@@ -71,6 +71,7 @@ pub mod subcommand_pipeline;
 pub mod subcommand_rtmp_inbound_endpoint;
 pub mod subcommand_rtmp_outbound_endpoint;
 pub mod subcommand_rtmp_publish;
+pub mod subcommand_server;
 pub mod subcommand_tune;
 pub mod subcommand_vmaf;
 pub mod tcp;